Tweak: Deletion of the Funky Emoji
Tweak: I deleted the emoji that was present in the script when the script started the parallel search.

I tried to play around with that but I felt that it looked too 'childish' in the release of the script. I accidently forgot to remove this one.

The emojis surrounding the 'Trusted' and the 'Untrusted' state remain to give the user a quick and clear answer.